# docs/swagger/collections.yaml
openapi: 3.0.0
servers:
  - description: Listmonk Developement Server
    url: http://localhost:9000/api
info:
  version: "1.0.0"
  title: Listmonk
  description: The API collection for listmonk
  license:
    name: AGPL-3.0 license
    url: https://github.com/knadh/listmonk/blob/master/LICENSE
...
paths:
...
  /lists:
    get:
      description: retrieves lists with additional metadata like subscriber counts. This may be slow.
      operationId: getLists
      tags:
        - Lists
      parameters:
        - in: query
          name: page
          description: total number of pages
          schema:
            type: integer
        - in: query
          name: per_page
          description: number of items per page
          schema:
            type: integer
      requestBody:
        required: true
        description: output parameters form
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                query:
                  description: Optional string to search a list by name.
                  type: string
                order_by:
                  description: Field to sort results by. name|status|created_at|updated_at
                  type: string
                order:
                  description: ASC|DESC Sort by ascending or descending order.
                  type: string
                minimal:
                  type: boolean
      responses:
        "200":
          description: list of metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: "#/components/schemas/List"
                      total:
                        type: integer
                      per_page:
                        type: integer
                      page:
                        type: integer

    post:
      description: handles list creation
      operationId: createList
      tags:
        - Lists
      requestBody:
        description: new list info
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewList"
      responses:
        "200":
          description: updated list object
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/List"
...
  "/public/subscription":
    post:
      description: handles subscription requests coming from public API calls.
      operationId: handlePublicSubscription
      tags:
        - Public
      requestBody:
        description: subscription request parameters
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                list_uuids:
                  type: array
                  items:
                    type: string

      responses:
        "200":
          description: response
          content:
            application/json:
              schema:
                type: object
                properties:
                  has_optin:
                    type: boolean
components:
  schemas:
...
    List:
      type: object
      properties:
        id:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string
        uuid:
          type: string
        name:
          type: string
        type:
          type: string
        optin:
          type: string
        tags:
          type: array
          items:
            type: string
        subscriber_count:
          type: integer
    NewList:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [public, private]
        optin:
          type: string
          enum: [single, double]
        tags:
          type: array
          items:
            type: string
...